Refactor + multi-sell + entities updates
Refactor some code in regards to the EntityExistance, so it can be used with Item and Entity since they share similar fields beside the core (item/entity) itself.
Also had to fix some bug where rotation of area items was not being applied due to the change of position and rotation (lights were facing the wrong way on load due to no rotation applied)
npc are no longer loaded on load, it's now constantly being updated to client so when they move the player will see them move
Implemented a function to allow selling multiple items or all items (this also include the transfer functionality)
Change a few models making itemObj to be entityObj so it can play well with the generic function of EntityExistance.